requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
hdbscan>=0.8.33
networkx>=3.1
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
openai>=1.0.0
pyvis>=0.3.2
plotly>=5.15
umap-learn>=0.5.3
matplotlib>=3.7
streamlit>=1.28.0
jupyter>=1.0.0
ipykernel>=6.0.0
node2vec>=0.4.6
